4 network topologies with same density for any number. results change!
"networks_equaldensity_stats-840_timing" shows the 4 networks and their stats (clustering, density, degree distribution). it is made to work with any number, works better (less spread in density between networks) with highly composite numbers. for 840 node points it takes about 30min to calculate, for 120 about 1 min.

in "simulation" i show how it can be implemented in a model. simulation is built up on "jose_iterationloop", so it runs with the original SIR modle as writen by Chris some time ago. the reason for that is that i really couldn't manage to make the new concept (using classes, as defined on "networkgen" work.

the only python code (not notebook" needed for both files is "functions_networks". it is called functions cause i define them as functions, not as classes.

there are many jpg screenshots showing how the results change depending on the network. name=network type, 240=number nodes, next number is p of changing opinion stochaicly, if there is another number it gets defined before, it is the "network specific" parameter. results are avaraged over "implementations=5".